Extend incorrect_fn_null_checks to incorrect_non_null_checks
This extends the scope of the incorrect_fn_null_checks lint to
also references.

In theory, we could extend the check even further,
for example to NonNull, or Box, but these types can't be casted
via as casts to raw pointers so it would be a bit harder from
an implementation point of view.