tests/ui/proc-macro/*: Migrate FIXMEs to check-pass
proc-macros are processed early in the compiler pipeline. There is no
need to involve codegen. So change to check-pass.

I have also looked through each changed test and to me it is
sufficiently clear that codegen is not needed for the purpose of the
test.

I skipped changing tests/ui/proc-macro/no-missing-docs.rs in this commit
because it was not clear to me that it can be changed to check-pass.