Update dist-various linux-gnu target baselines
These Tier 2 non-host targets were affected by the update to
`ubuntu:22.04` in #113931, but the platform support page was already out
of date after #81521 updated to `ubuntu:20.04`. They are now listed for
kernel 5.15 and glibc 2.35 as found in 22.04, and a new footnote
mentions that this only matters when using the dynamic `libstd.so`.