Add invisible open delimiters to `TokenDescription`.
Invisible delimiters pretty-print as empty strings, and changing that
can break some proc macros. But error messages saying "expected
identifer, found ``" are bad. So this commit adds support for invisible
open delimiters in `TokenDescription` so they print as "invisible open
delimiter" in error messages, instead of "``".

It's not used meaningfully yet, but will be needed to get rid of interpolated
tokens.