Reference uplifted clippy lints' rustc name in the release notes
I updated to Rust 1.72.0, got a new warning from rustc, wanted to read
more about it, so I went to RELEASES.md and searched for the new lint's
name as shown in the warning I got.

I found no results because the relevant entry in RELEASES only contained
the lint's old Clippy name, not its new rustc name.

This adds the rustc name for lints so that someone doing the same thing
I did will have more success.

Some of the uplifted lints didn't have a name change, so I didn't add
the rustc name as a search for it will succeed.