compiler/hir/check: clarify error message on transparent ADT
Replace the phrase "layout of non-zero size or alignment larger than 1"
with "non-trivial layout", and then elaborate that a layout is trivial
if, and only if, it is zero-sized with an alignment of 1.

Additionally, add a comment to the check-function explaining which
information is gathered to check transparent ADTs, and elaborate why it
is needed.